#71: 'Space' hotkey for play/pause
This was a bit tricky because space is the default key for firing the
focussed JavaFx node. So I created subclasses (Unfocusable*) for the
relevant elements which never request focus and pass it on to the root
node instead. In cominbation with a mouse-click handler for the scene
(to remove focus from the comment text area), this approach seems to
work pretty good.