#81: Progress overlay while adding tracks
This required to wrap the root layout into a StackPane. Because tracks
are now added on a background thread, every action that interacts with
the UI needs to be wrapped in JavaFX Platform.runLater(...).
Maybe there's also a better solution but right now I hope I got all
method-calls that required modification.