#65: Postpone opening of clip after audio analysis
By storing the number of channels and the frame length in the track
class, the analysis step of the constructor no needs the clip. This
means that the clip can be opened at the end of the constructor, when
the temporary arrays of the audio analysis are already ready for garbage
collection. (TL;DR: can open 4x10min WAV now :))